renamed NewSidebar to Sidebar
